start SessionPair file
